+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 ### Fixed
 - Fixed search bar issues in workshops.
+- Fixed issues in register workshops
 - Fixed error messajes in register beneficiary
 
 ## [0.1.0] - 18/11/2025